# swagger-api-v3.json
{
  "swagger": "2.0",
  "info": {
    "description": "本文档说明了每个接口返回、接收的字段信息。并且制定相应的规范作为前后端交流的渠道。",
    "version": "2.0",
    "title": "易授微服务RESTful APIs说明文档",
    "contact": {
      "name": "刘祥德",
      "url": "/",
      "email": "luvletterU@njust.edu.cn"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
...
    "/arrangements/{arrangementId}/tests": {
      "get": {
        "tags": [
          "teacher-arrangement-controller"
        ],
        "summary": "获取给该门课下发起的所有测试",
        "operationId": "findAllOnlineClassTestUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "isOpening",
            "in": "query",
            "description": "isOpening",
            "required": false,
            "type": "boolean",
            "default": true
          },
          {
            "name": "isTimeOut",
            "in": "query",
            "description": "isTimeOut",
            "required": false,
            "type": "boolean",
            "default": false
          },
          {
            "name": "isFinished",
            "in": "query",
            "description": "isFinished",
            "required": false,
            "type": "boolean",
            "default": false
          },
          {
            "name": "arrangementId",
            "in": "path",
            "description": "教师安排表的编号",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/OnlineClassTest"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "teacher-arrangement-controller"
        ],
        "summary": "教师创建一个在线测试,此测试跟对应的教师安排相关联",
        "operationId": "createOnlineClassTestUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "onlineClassTest",
            "description": "onlineClassTest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OnlineClassTest"
            }
          },
          {
            "name": "arrangementId",
            "in": "path",
            "description": "教师安排表的编号",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/OnlineClassTest"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/schedules/{scheduleId}/test_records": {
      "get": {
        "tags": [
          "class-schedule-controller"
        ],
        "summary": "findTestRecords",
        "operationId": "findTestRecordsUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "scheduleId",
            "in": "path",
            "description": "scheduleId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TestRecord"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/sign_records/{signId}/schedule": {
      "get": {
        "tags": [
          "sign-record-controller"
        ],
        "summary": "findClassSchedule",
        "operationId": "findClassScheduleUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "signId",
            "in": "path",
            "description": "signId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ClassSchedule"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/test_records": {
      "get": {
        "tags": [
          "test-record-controller"
        ],
        "summary": "findAllTestRecords",
        "operationId": "findAllTestRecordsUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TestRecord"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "test-record-controller"
        ],
        "summary": "updateTestRecord",
        "operationId": "updateTestRecordUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "testRecord",
            "description": "testRecord",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TestRecord"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TestRecord"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/test_records/{recordId}": {
      "post": {
        "tags": [
          "test-record-controller"
        ],
        "summary": "createTestRecord",
        "operationId": "createTestRecordUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "testRecord",
            "description": "testRecord",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TestRecord"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TestRecord"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/test_records/{recordId}/schedule": {
      "get": {
        "tags": [
          "test-record-controller"
        ],
        "summary": "findClassSchedule",
        "operationId": "findClassScheduleUsingGET_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "recordId",
            "in": "path",
            "description": "recordId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ClassSchedule"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/test_records/{recordId}/test": {
      "get": {
        "tags": [
          "test-record-controller"
        ],
        "summary": "findTest",
        "operationId": "findTestUsingGET_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "recordId",
            "in": "path",
            "description": "recordId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/OnlineClassTest"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/test_records/{testId}": {
      "delete": {
        "tags": [
          "test-record-controller"
        ],
        "summary": "deleteTestRecord",
        "operationId": "deleteTestRecordUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "testId",
            "in": "path",
            "description": "testId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TestRecord"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/tests": {
      "get": {
        "tags": [
          "test-controller"
        ],
        "summary": "findAllTests",
        "operationId": "findAllTestsUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/OnlineClassTest"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "test-controller"
        ],
        "summary": "createTest",
        "operationId": "createTestUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "onlineClassTest",
            "description": "onlineClassTest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OnlineClassTest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/OnlineClassTest"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "test-controller"
        ],
        "summary": "updateTest",
        "operationId": "updateTestUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "onlineClassTest",
            "description": "onlineClassTest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OnlineClassTest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/OnlineClassTest"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/tests/{testId}": {
      "get": {
        "tags": [
          "test-controller"
        ],
        "summary": "findTest",
        "operationId": "findTestUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "testId",
            "in": "path",
            "description": "testId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/OnlineClassTest"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "test-controller"
        ],
        "summary": "deleteTest",
        "operationId": "deleteTestUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "testId",
            "in": "path",
            "description": "testId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/OnlineClassTest"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ClassSchedule": {
      "type": "object",
      "properties": {
        "courseComment": {
          "$ref": "#/definitions/CourseComment"
        },
        "scheduleId": {
          "type": "integer",
          "format": "int32"
        },
        "student": {
          "$ref": "#/definitions/Student"
        },
        "teacherArrangement": {
          "$ref": "#/definitions/TeacherArrangement"
        }
      }
    },
...
    "Course": {
      "type": "object",
      "properties": {
        "courseId": {
          "type": "integer",
          "format": "int32"
        },
        "courseName": {
          "type": "string"
        },
        "credit": {
          "type": "number",
          "format": "double"
        },
        "type": {
          "type": "string"
        }
      }
    },
    "CourseComment": {
      "type": "object",
      "properties": {
        "arrangement": {
          "$ref": "#/definitions/TeacherArrangement"
        },
        "commentId": {
          "type": "integer",
          "format": "int32"
        },
        "schedule": {
          "$ref": "#/definitions/ClassSchedule"
        },
        "score": {
          "type": "number",
          "format": "float"
        }
      }
    },
...
    "Department": {
      "type": "object",
      "properties": {
        "depId": {
          "type": "integer",
          "format": "int32"
        },
        "depName": {
          "type": "string"
        }
      }
    },
...
    "OnlineClassTest": {
      "type": "object",
      "properties": {
        "arrangement": {
          "$ref": "#/definitions/TeacherArrangement"
        },
        "deadline": {
          "type": "string",
          "format": "date-time"
        },
        "establishedTime": {
          "type": "string",
          "format": "date-time"
        },
        "questions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Question"
          }
        },
        "testId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "Question": {
      "type": "object",
      "properties": {
        "answerOrder": {
          "type": "integer",
          "format": "int32"
        },
        "content": {
          "type": "string"
        },
        "course": {
          "$ref": "#/definitions/Course"
        },
        "options": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/QuestionOption"
          }
        },
        "questionId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "QuestionOption": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        },
        "optionId": {
          "type": "integer",
          "format": "int32"
        },
        "optionOrder": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "Student": {
      "type": "object",
      "properties": {
        "department": {
          "$ref": "#/definitions/Department"
        },
        "grade": {
          "type": "string"
        },
        "majorIn": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "studentGender": {
          "type": "string"
        },
        "studentId": {
          "type": "string"
        },
        "studentName": {
          "type": "string"
        },
        "userId": {
          "type": "integer",
          "format": "int32"
        },
        "username": {
          "type": "string"
        }
      }
    },
    "Teacher": {
      "type": "object",
      "properties": {
        "department": {
          "$ref": "#/definitions/Department"
        },
        "password": {
          "type": "string"
        },
        "teacherGender": {
          "type": "string"
        },
        "teacherId": {
          "type": "string"
        },
        "teacherName": {
          "type": "string"
        },
        "userId": {
          "type": "integer",
          "format": "int32"
        },
        "username": {
          "type": "string"
        }
      },
      "description": "教师实体"
    },
    "TeacherArrangement": {
      "type": "object",
      "properties": {
        "arrangementId": {
          "type": "integer",
          "format": "int32"
        },
        "course": {
          "$ref": "#/definitions/Course"
        },
        "place": {
          "type": "string"
        },
        "teacher": {
          "$ref": "#/definitions/Teacher"
        }
      }
    },
    "TestRecord": {
      "type": "object",
      "properties": {
        "answer": {
          "type": "string"
        },
        "recordId": {
          "type": "integer",
          "format": "int32"
        },
        "schedule": {
          "$ref": "#/definitions/ClassSchedule"
        },
        "score": {
          "type": "number",
          "format": "float"
        },
        "test": {
          "$ref": "#/definitions/OnlineClassTest"
        }
      }
    },
...
  }
}